Create README.md
To introduce myself, I was a Post graduate in Human Resource Management with 7 years of Industry experience in people management. Off late I have observed trends in the industry are changing with more of algorithms and bots handling the routine tasks, which has eased the way clients are handled and business processes are delivered with better satisfaction quotient.
The Human Resource and recruitment industry is not an exception to the transformation being witnessed in the automation space handling stake holder management and reducing turnaround times in various processes of the Human resource life cycle in the corporate space.
This has made me turn my attention towards data science and machine learning, which if effectively implemented will make transformational impacts in any industry by unveiling patterns from the data existing with various sources and training the systems to automate to the best possible extent to make the tasks easy and performed with little or no human intervention.
My experience handling people and various business circumstances in the industry and my background in information science gave me an effective platform to launch in the data science space and to hone my skills with my academic sessions at the elite Royal Holloway University of London. Together with my analytical and the recently acquired academic knowledge in the stream of Data Analytics, I am foraging to launch into the industry to uncover the hidden potential of data to the benefit of the business and society at large.
I am looking for a placement in the Industry to work with real world complexities of data handling and management to realise my potential as a future data scientist.